flask/requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
Werkzeug==0.11.4
itsdangerous==0.24
